address bioc review (Herve)
I addressed all the issues you pointed out. The new tarball is uploaded
to the bioc tracker ("les_0.13.1.tar.gz"). It is tested with R-2.11.1
and R-2.12.0dev on a linux system.
I could not reproduce issue 5 and the fix for this is not verified.
Could you please check this issue and inform me if it still occurs?
Further issue 11 is still open, but it will require more time for a good
solution.
All the other issues should be fixed. If there is anything else, please
let me know.

> 1) 'R CMD check' WARNING:
>
>    * checking for missing documentation entries ... WARNING
>    Undocumented data sets:
>      borders chr pos pval
>    All user-level objects in a package should have documentation entries.
>    See the chapter 'Writing R documentation files' in manual 'Writing R
>    Extensions'.
>
added alias to statTile.Rd
This fixes the warning for R-2.12.0, for R-2.11.1 the warning does not
occur.

> 2) In DESCRIPTION file:
>
>      The package provides a universal framework suitable of identifying...
>                                                 ^^^^^^^^^^^
>    "suitable for" or "capable of" (or just "for")
>
corrected

> 3) In the man page of estimate():
>
>      This can be useful on multi-core machines for large datasets and
>      intensive CI computations.
>                ^^
>
>    Probably the result of a plain copy/paste from the ci() man page.
>
text adapted in estimate.Rd and ci.Rd

> 4) Why put the examples in a \dontrun{} statement? Keeping the
>    code using multicore in a \dontrun{} statement is OK though.
>
I removed the dontrun for most examples. Any multicore and export()
calls are still in dontrun statements. The examples of export() write
data to the disk which I'd like to avoid. If there is an elegant
solution in preventing functions writing to disk please let me know.

> 5) In man page for simTile:
>
>      Simulated tiling array data set containing Loci of Enhanced
>
>    Shouldn't it be "Loci of Enhanced Significance"?
>
I could not reproduce this with R-2.11 and R-2-12; in the man pages the
full title "Simulated tiling array data set containing Loci of Enhanced
Significance" is shown.
This might have been caused by a line break in the title of simTile.Rd
which I fixed. Could you please verify if the title is displayed
correctly now?

> 6) Using "multicore" %in% .packages() is supposed to be THE way instead
>    of using any(.packages() %in% "multicore").
>
Yes, that is the better solution. Corrected in allInteral.R.

> 7) Your create() function is a constructor for Les objects. If all
>    constructors in Bioconductor were called create() or construct()
>    we would have hundreds of create() functions that would conflict
>    with each other (think of the user that has a dozen or more packages
>    loaded in his/her session, each of them defining a create() function).
>    Also what name will you use the day you want to add another class
>    to your package and provide a constructor for it?
>    My point is that the name "create" is too general. The consensus in
>    Bioconductor is to name the constructor after the class (this is
>    enforced in other languages). So you could use Les(). But also
>    something like createLes() would be OK (still better than create()).
>
changed to "Les()" in code, documentation and NAMESPACE

> 8) Using threshold() on a Les object returned by create():
>
>      > x <- create(pos, pval, chr)
>      > threshold(x, verbose=TRUE)
>      [1] "88 significant probes estimated with limit Lambda>=NA"
>      ** Object of class 'Les' **
>      * 1000 probes on 2 chromosomes
>      * 89 regulated probes estimated for lambda >= NA
>
>    is the NA expected?
>
It is expected, however I agree that it may be confusing. I adapted the
code of the "show" method in allMethods.R to avoid it.
Now "for lambda >= NA" is omitted if it is not informative.

> 9) In man page for ci():
>
>      \item{object}{Object of class 'Les' as returned by 'estimate' or
>        subsequent methods.}
>
>    What are the "subsequent methods"?
>
corrected in ci.Rd

> 10) Using create(), followed by estimate(), followed by ci() seems to
>    be the typical workflow. Each step adds data to the Les object.
>    Unlike the examples in the package that use 'x' for input and output,
>    I've used names 'x', 'y', 'z':
>
>      > x
>      ** Object of class 'Les' **
>      * 1000 probes on 2 chromosomes
>      > y
>      ** Object of class 'Les' **
>      * 1000 probes on 2 chromosomes
>      * Lambda in range [0, 0.994175] with window size 25
>      > z
>      ** Object of class 'Les' **
>      * 1000 probes on 2 chromosomes
>      * Lambda in range [0, 0.994175] with window size 25
>      * 0.95 confidence intervals computed for 100 probes
>
>    I can't pass 'x' directly to ci() (it's a Les object but it would
>    need to go thru estimate() first):
>
>      > ci(x, subset=1:100, conf=0.95, nCores=2)
>      Error in from:to : argument of length 0
>
>    The error message would need to be more informative.
updated the error message in allMethods.R, based on checkState() (see
issue 11)

>
>    What happens if I call estimate() on 'z'? It seems to work but do
>    I get a valid object?
>
The data itself is valid, but it may get mixed up if one calls
estimate() with different parameters.
I corrected it in estimate() by overwriting fields that already contain
data from ci(), and added a note in the estimate man.

> 11) So it's like a Les object can have different "states": its current
>     state determines what transformations can be applied to it and each
>     transformation puts the object in a new state.
>     I don't know if this "state" is stored somewhere in the object.
>     If not, it can probably be inferred from the content of the slots.
>     Maybe an experienced user will be able to guess it based on
>     what's displayed by the show() method.
>     Anyway, maybe there should be a robust mechanism in place to ensure
>     that estimate(), threshold(), regions() and ci() are called on a
>     Les object with the right state and return an Les object with the
>     new state. There are 2 fairly standard aproaches for doing this:
>
>       * One is to use a slot for storing the state (aka "flag").
>
>       * The other one is to use several classes. For example create()
>         returns class ALes, estimate() is a method for ALes objects
>         that returns a BLes object and ci() is a method for BLes objects
>         that returns a CLes object (of course you want to use better names).
>         Each class contains the previous class and adds slots to it.
>         With this approach, in ci() man page, you can replace:
>
>           \item{object}{Object of class 'Les' as returned by 'estimate' or
>           subsequent methods.}
>
>         with
>
>           \item{object}{Object of class 'BLes'.}
>
>         and
>
>           \value{
>             Object of class 'Les' with additionally filled slots:
>             ci, subset, nBoot, conf
>           }
>
>         width
>
>           \value{
>             Object of class 'CLes'.
>           }
>
Yes, the different states of the "Les" object are an issue. The problem
is that the analysis does not have to be strictly linear and should be
adapted in specific situation. I don't have a clever solution for this
and have to think a little more about this issue.
For the moment I follow your suggestion with assigning states to the
object. However I will have to put more work into the code as well as
documentation, and the implementation with setState() and checkState()
is just a first step.